Remove duplicate queue code from data worker
