+++ apps/data-worker/tsconfig.json
@@ -7,7 +7,7 @@
       "@otr/core": ["../../packages/otr-core/src/index.ts"],
       "@otr/core/*": ["../../packages/otr-core/src/*"],
       "@otr/core/osu": ["../../packages/otr-core/src/osu/enums.ts"],
-      "@otr/core/queues": ["../../packages/otr-core/src/queues/constants.ts"]
+      "@otr/core/queues": ["../../packages/otr-core/src/queues/index.ts"]
     }
   },
   "include": ["env.d.ts", "src/**/*.ts"],
